backend/requirements-docker.txt
ariadne==0.21
asyncio==3.4.3
asyncpg==0.29.0
fastapi==0.109.1
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.7
python-jose==3.3.0
uvicorn==0.25.0
